IdmResource tests suite is created
