+++ src/test/java/gov/ca/cwds/idm/IdmResourceTestSuite.java
@@ -1,0 +1,13 @@
+package gov.ca.cwds.idm;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+
+@RunWith(Suite.class)
+@Suite.SuiteClasses({
+    IdmResourceTest1.class,
+    IdmResourceTest2.class
+})
+public class IdmResourceTestSuite {
+
+}
